lapcounter-server-gpio/requirements.txt
RPi.GPIO==0.7.1
paho-mqtt==1.6.1
